+++ clients/wallet/src/main/java/com/binance/connector/client/wallet/rest/model/DustConvertRequest.java
@@ -43,6 +43,12 @@ public class DustConvertRequest {
     @jakarta.annotation.Nonnull
     private String asset;
 
+    public static final String SERIALIZED_NAME_ACCOUNT_TYPE = "accountType";
+
+    @SerializedName(SERIALIZED_NAME_ACCOUNT_TYPE)
+    @jakarta.annotation.Nullable
+    private String accountType;
+
     public static final String SERIALIZED_NAME_CLIENT_ID = "clientId";
 
     @SerializedName(SERIALIZED_NAME_CLIENT_ID)
@@ -88,6 +94,25 @@ public class DustConvertRequest {
 
     public void setAsset(@jakarta.annotation.Nonnull String asset) {
         this.asset = asset;
+    }
+
+    public DustConvertRequest accountType(@jakarta.annotation.Nullable String accountType) {
+        this.accountType = accountType;
+        return this;
+    }
+
+    /**
+     * Get accountType
+     *
+     * @return accountType
+     */
+    @jakarta.annotation.Nullable
+    public String getAccountType() {
+        return accountType;
+    }
+
+    public void setAccountType(@jakarta.annotation.Nullable String accountType) {
+        this.accountType = accountType;
     }
 
     public DustConvertRequest clientId(@jakarta.annotation.Nullable String clientId) {
@@ -180,6 +205,7 @@ public class DustConvertRequest {
         }
         DustConvertRequest dustConvertRequest = (DustConvertRequest) o;
         return Objects.equals(this.asset, dustConvertRequest.asset)
+                && Objects.equals(this.accountType, dustConvertRequest.accountType)
                 && Objects.equals(this.clientId, dustConvertRequest.clientId)
                 && Objects.equals(this.targetAsset, dustConvertRequest.targetAsset)
                 && Objects.equals(this.thirdPartyClientId, dustConvertRequest.thirdPartyClientId)
@@ -191,7 +217,12 @@ public class DustConvertRequest {
     @Override
     public int hashCode() {
         return Objects.hash(
-                asset, clientId, targetAsset, thirdPartyClientId, dustQuotaAssetToTargetAssetPrice);
+                asset,
+                accountType,
+                clientId,
+                targetAsset,
+                thirdPartyClientId,
+                dustQuotaAssetToTargetAssetPrice);
     }
 
     @Override
@@ -199,6 +230,7 @@ public class DustConvertRequest {
         StringBuilder sb = new StringBuilder();
         sb.append("class DustConvertRequest {\n");
         sb.append("		asset: ").append(toIndentedString(asset)).append("\n");
+        sb.append("		accountType: ").append(toIndentedString(accountType)).append("\n");
         sb.append("		clientId: ").append(toIndentedString(clientId)).append("\n");
         sb.append("		targetAsset: ").append(toIndentedString(targetAsset)).append("\n");
         sb.append("		thirdPartyClientId: ")
@@ -218,6 +250,10 @@ public class DustConvertRequest {
         String assetValueAsString = "";
         assetValueAsString = assetValue.toString();
         sb.append("asset=").append(urlEncode(assetValueAsString)).append("");
+        Object accountTypeValue = getAccountType();
+        String accountTypeValueAsString = "";
+        accountTypeValueAsString = accountTypeValue.toString();
+        sb.append("accountType=").append(urlEncode(accountTypeValueAsString)).append("");
         Object clientIdValue = getClientId();
         String clientIdValueAsString = "";
         clientIdValueAsString = clientIdValue.toString();
@@ -268,6 +304,7 @@ public class DustConvertRequest {
         // a set of all properties/fields (JSON key names)
         openapiFields = new HashSet<String>();
         openapiFields.add("asset");
+        openapiFields.add("accountType");
         openapiFields.add("clientId");
         openapiFields.add("targetAsset");
         openapiFields.add("thirdPartyClientId");
@@ -312,6 +349,14 @@ public class DustConvertRequest {
                             "Expected the field `asset` to be a primitive type in the JSON string"
                                     + " but got `%s`",
                             jsonObj.get("asset").toString()));
+        }
+        if ((jsonObj.get("accountType") != null && !jsonObj.get("accountType").isJsonNull())
+                && !jsonObj.get("accountType").isJsonPrimitive()) {
+            throw new IllegalArgumentException(
+                    String.format(
+                            "Expected the field `accountType` to be a primitive type in the JSON"
+                                    + " string but got `%s`",
+                            jsonObj.get("accountType").toString()));
         }
         if ((jsonObj.get("clientId") != null && !jsonObj.get("clientId").isJsonNull())
                 && !jsonObj.get("clientId").isJsonPrimitive()) {